spinnaker_links are spin1 only
