+++ spinn_machine/version/abstract_version.py
@@ -403,18 +403,6 @@ class AbstractVersion(object, metaclass=AbstractBase):
         raise NotImplementedError
 
     @abstractmethod
-    def spinnaker_links(self) -> List[Tuple[int, int, int]]:
-        """
-        The list of Local X, Y and link Id to add spinnaker links to
-
-        These are applied local to each Ethernet Chip and only if the link is
-        not connected to another board
-
-        :returns: List of Tuples of Local X, Y, and spinnaker link Id
-        """
-        raise NotImplementedError
-
-    @abstractmethod
     def quads_maps(self) -> Optional[Dict[int, Tuple[int, int, int]]]:
         """
         If applicable returns a map of virtual id to quad qx, qy, qp
